backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart
httpx
pymilvus
redis
celery
python-dotenv
pypdf
langchain-community
sentence-transformers
